+++ ItsGrey/data/book/src/main/java/com/tntt/book/repository/BookRepositoryImpl.kt
@@ -34,7 +34,7 @@ class BookRepositoryImpl @Inject constructor(
         }
     }
 
-    override suspend fun getBookInfoList(userId: String, sortType: SortType, startIndex: Long, bookType: BookType): Flow<List<BookInfo>> = flow {
+    override suspend fun getBookInfoList(userId: String, sortType: SortType, startIndex: Int, bookType: BookType): Flow<List<BookInfo>> = flow {
         bookDataSource.getBookDtoList(userId, sortType, startIndex, bookType).collect() { bookDtoList ->
             val bookInfoList = mutableListOf<BookInfo>()
             for (bookDto in bookDtoList){
